refactor: restructure build process by consolidating Rollup configurations and modularizing files
