+++ src/commands/dev.js
@@ -4,7 +4,7 @@ import {
   withBuildOptions,
   withServerOptions,
 } from "#commands/_sharedOptions.js";
-import { buildWithRollup, cleanupDist } from "#scripts/build/rollup.js";
+import { buildWithRollup, cleanupDist } from "#scripts/build/index.js";
 
 let devCommand = program
   .command("dev")
